requirements.txt
flask
Flask-RESTful
python-dotenv
psycopg2-binary
SQLAlchemy
flask_sqlalchemy
Flask-Migrate
flask-cors